+++ core/modules/modBulkaddproducts.class.php
@@ -47,20 +47,20 @@ class modBulkaddproducts extends DolibarrModules
         $this->descriptionlong = "This module provides hook-based functionality to add multiple products to customer and supplier documents including Quotations, Sales Orders, Customer Invoices, Purchase Orders and Supplier Invoices.";
 
         // Possible values for version are: 'development', 'experimental', 'dolibarr', 'dolibarr_deprecated' or a version string like 'x.y.z'
-        $this->version = '1.0.4';
-        
+        $this->version = '1.0.5';
+
         // Url to the file with your last numberversion of this module
         $this->url_last_version = '';
-        
+
         // Key used in llx_const table to save module status enabled/disabled (where MYMODULE is value of property name of module in uppercase)
         $this->const_name = 'MAIN_MODULE_'.strtoupper($this->name);
-        
+
         // Name of image file used for this module.
         $this->picto = 'generic';
 
         // Define some features supported by module (triggers, login, substitutions, menus, css, etc...)
         $this->module_parts = array(
-            'hooks' => array('ordercard', 'orderline', 'propalcard', 'propalline', 'invoicecard', 'invoiceline', 'supplier_proposalcard', 'supplier_proposalline', 'ordersuppliercard', 'ordersupplierline', 'invoicesuppliercard', 'invoicesupplierline'),
+            'hooks' => array('ordercard', 'orderline', 'propalcard', 'propalline', 'invoicecard', 'invoiceline', 'supplier_proposalcard', 'supplier_proposalline', 'ordersuppliercard', 'ordersupplierline', 'supplier_ordercard', 'supplier_orderline', 'invoicesuppliercard', 'invoicesupplierline'),
             'triggers' => 0,
             'login' => 0,
             'substitutions' => 0,